+++ Changelog.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Changed
+- Update _Android Gradle Plugin_ to 4.0.0, see [release notes](https://developer.android.com/studio/releases/gradle-plugin)
+- Update _Detekt_ to version _1.9.1_ see [release notes](https://github.com/detekt/detekt/releases/tag/v1.9.1)
 
 ## [0.13.0] - 2020-05-19
 ## Fixed